% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{clean_accounts_df}
\alias{clean_accounts_df}
\title{clean raw accounts df}
\usage{
clean_accounts_df(accounts_df)
}
\arguments{
\item{accounts_df}{raw mint accounts df}
}
\description{
clean raw accounts df
}
