% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{clean_transactions_df}
\alias{clean_transactions_df}
\title{clean raw transactions df}
\usage{
clean_transactions_df(transactions_df)
}
\arguments{
\item{transactions_df}{raw mint transactions df}
}
\description{
clean raw transactions df
}
