% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mint_data.R
\name{get_current_projections}
\alias{get_current_projections}
\title{Get current financial account projections}
\usage{
get_current_projections(
  historical_transactions_df,
  accounts_df,
  config_list,
  forecast_date_range
)
}
\arguments{
\item{historical_transactions_df}{aggregated and cleaned transactions}

\item{accounts_df}{aggregated and cleaned accounts}

\item{config_list}{user config data}

\item{forecast_date_range}{tuple of start and end date}
}
\description{
Get current financial account projections
}
