% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_utils.R
\name{get_manual_adjustments}
\alias{get_manual_adjustments}
\title{Extract manual account adjustments from config}
\usage{
get_manual_adjustments(config_file)
}
\arguments{
\item{config_file}{user json config}
}
\value{
tibble of adjustments
}
\description{
Extract manual account adjustments from config
}
