% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mint_data.R
\name{get_mint_data_by_type_memoised}
\alias{get_mint_data_by_type_memoised}
\title{Get mint data by type memoised}
\usage{
get_mint_data_by_type_memoised(data_name, user_name, read_cache, write_cache)
}
\arguments{
\item{data_name}{name of data to get in switch statement}

\item{user_name}{user name for config file}

\item{read_cache}{bool}

\item{write_cache}{bool}
}
\description{
Get mint data by type memoised
}
