% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_projections}
\alias{get_projections}
\title{Create financial projections based on past data and manual adjustments}
\usage{
get_projections(
  transactions,
  account_df,
  config_file,
  forecast_date_range,
  historical_start_date
)
}
\arguments{
\item{transactions}{df of transactions from mint by category}

\item{account_df}{df of accounts from mint}

\item{config_file}{user config file}

\item{forecast_date_range}{date range to forecast}

\item{historical_start_date}{start of historical data}
}
\description{
Create financial projections based on past data and manual adjustments
}
