% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator_utils.R
\name{summarise_categories}
\alias{summarise_categories}
\title{Get transactions grouped by category}
\usage{
summarise_categories(
  transactions,
  config_file,
  start_date,
  agg_vec,
  include_outlier = FALSE
)
}
\arguments{
\item{transactions}{df of mint transactions}

\item{config_file}{user config file}

\item{start_date}{date to keep transactions after}

\item{agg_vec}{vector for group_by columns}

\item{include_outlier}{whether to remove outliers as defined in user config}
}
\description{
Get transactions grouped by category
}
