% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mint_data.R
\name{summarize_transactions}
\alias{summarize_transactions}
\title{Get mint historical data aggregated}
\usage{
summarize_transactions(
  transactions_df,
  time_vec,
  config_file,
  historical_start_date
)
}
\arguments{
\item{transactions_df}{raw transactions}

\item{time_vec}{vector to aggregate by}

\item{config_file}{user config file}

\item{historical_start_date}{drop before this date}
}
\description{
Get mint historical data aggregated
}
